Start refactoring to make SyntaxHighlightReconciler more testable
