+++ browser/test/Services/SyntaxHighlighting/SyntaxHighlightingReconcilerTests.ts
@@ -1,0 +1,9 @@
+import * as assert from "assert"
+
+// import * as SyntaxHighlightReconciler from "./../../../src/Services/SyntaxHighlighting/SyntaxHighlightReconciler"
+
+describe("SyntaxHighlightReconciler", () => {
+    it("handles case where buffer has no syntax info", () => {
+        assert.ok(false)
+    })
+})
